refactor(ci,services,app_sources): consolidate error handling and foreground service logic
- Simplify error handling across app sources by replacing verbose try-catch blocks with empty catches
- Move foreground service initialization from main.dart to bg_updates.dart with state tracking
- Update GitGuardian action version from v1.51.0 to main branch
- Replace hardcoded Colors.red with Theme.of(context).colorScheme.error in settings page
- Comment out inactive locale entries and remove unused imports across the codebase
