+++ lib/app_sources/gitea.dart
@@ -223,9 +223,7 @@ class Gitea extends AppSource {
     if (createdAt != null) {
       try {
         releaseDate = DateTime.parse(createdAt as String);
-      } catch (_) {
-        // Ignore parse errors
-      }
+      } catch (_) {}
     }
 
     String? changeLog = targetRelease['body'] as String?;
